Renamed some packages to the java code style. Fixed some code-style issues
