+++ src/META-INF/MANIFEST.MF
@@ -1,4 +1,4 @@
 Manifest-Version: 1.0
-Main-Class: com.killersssurprise.Convertor.SimpleImageToPixelbattleIma
+Main-Class: com.killersssurprise.сonverter.SimpleImageToPixelbattleIma
  geConvector
